+++ src/sai.rs
@@ -142,20 +142,25 @@ pub struct Asynchronous;
 pub struct AsyncMaster<Ch: SaiChannel> {
     // TODO: Remove attribute when destroy function is implemented.
     #[allow(unused)]
-    pins: (Ch::Mclk, Ch::Fs, Ch::Sck, Ch::Sd),
+    pins: (Option<Ch::Mclk>, Ch::Fs, Ch::Sck, Ch::Sd),
 }
 
 impl<Ch: SaiChannel> AsyncMaster<Ch> {
     fn new(
         pins: (
-            impl Into<Ch::Mclk>,
+            Option<impl Into<Ch::Mclk>>,
             impl Into<Ch::Fs>,
             impl Into<Ch::Sck>,
             impl Into<Ch::Sd>,
         ),
     ) -> Self {
         Self {
-            pins: (pins.0.into(), pins.1.into(), pins.2.into(), pins.3.into()),
+            pins: (
+                pins.0.map(Into::into),
+                pins.1.into(),
+                pins.2.into(),
+                pins.3.into(),
+            ),
         }
     }
 }
@@ -164,20 +169,25 @@ impl<Ch: SaiChannel> AsyncMaster<Ch> {
 pub struct AsyncSlave<Ch: SaiChannel> {
     // TODO: Remove attribute when destroy function is implemented.
     #[allow(unused)]
-    pins: (Ch::Mclk, Ch::Fs, Ch::Sck, Ch::Sd),
+    pins: (Option<Ch::Mclk>, Ch::Fs, Ch::Sck, Ch::Sd),
 }
 
 impl<Ch: SaiChannel> AsyncSlave<Ch> {
     fn new(
         pins: (
-            impl Into<Ch::Mclk>,
+            Option<impl Into<Ch::Mclk>>,
             impl Into<Ch::Fs>,
             impl Into<Ch::Sck>,
             impl Into<Ch::Sd>,
         ),
     ) -> Self {
         Self {
-            pins: (pins.0.into(), pins.1.into(), pins.2.into(), pins.3.into()),
+            pins: (
+                pins.0.map(Into::into),
+                pins.1.into(),
+                pins.2.into(),
+                pins.3.into(),
+            ),
         }
     }
 }
@@ -609,7 +619,7 @@ where
     pub fn master_rx(
         self,
         pins: (
-            impl Into<Ch::Mclk>,
+            Option<impl Into<Ch::Mclk>>,
             impl Into<Ch::Fs>,
             impl Into<Ch::Sck>,
             impl Into<Ch::Sd>,
@@ -633,7 +643,7 @@ where
     pub fn master_tx(
         self,
         pins: (
-            impl Into<Ch::Mclk>,
+            Option<impl Into<Ch::Mclk>>,
             impl Into<Ch::Fs>,
             impl Into<Ch::Sck>,
             impl Into<Ch::Sd>,
@@ -657,7 +667,7 @@ where
     pub fn slave_rx(
         self,
         pins: (
-            impl Into<Ch::Mclk>,
+            Option<impl Into<Ch::Mclk>>,
             impl Into<Ch::Fs>,
             impl Into<Ch::Sck>,
             impl Into<Ch::Sd>,
@@ -678,7 +688,7 @@ where
     pub fn slave_tx(
         self,
         pins: (
-            impl Into<Ch::Mclk>,
+            Option<impl Into<Ch::Mclk>>,
             impl Into<Ch::Fs>,
             impl Into<Ch::Sck>,
             impl Into<Ch::Sd>,